+++ plugins/pluginLib/canvas/canvas.cpp
@@ -26,14 +26,14 @@ void copyImageToLayer(ILayer* dst, const IImage* src, const vec2i& layerPos, con
 {
     assert(src->getSize().x == 1920 && src->getSize().y == 1080);
 
-    for (size_t x = 0; x < size.x; ++x)
+    for (unsigned x = 0; x < size.x; ++x)
     {
-        for (size_t y = 0; y < size.y; ++y)
+        for (unsigned y = 0; y < size.y; ++y)
         {
             if (src->getPixel(x, y).a == 0)
                 continue;
 
-            dst->setPixel(vec2i{x, y} - layerPos, src->getPixel(x, y));
+            dst->setPixel(vec2i{static_cast<int>(x), static_cast<int>(y)} - layerPos, src->getPixel(x, y));
         }
     }
 }
